+++ src/decorators/index.ts
@@ -1,3 +1,4 @@
 export * from './dependencies';
+export * from './di';
 export * from './inject';
 export * from './singleton';