+++ src/decorators/singleton.ts
@@ -1,3 +1,4 @@
+import { di } from './di';
 import { InheritancePreserver } from './InheritancePreserver';
 import { singletonSymbol as singleton_s } from '../symbols';
 import { Constructor, ImplementationType } from '../types';
@@ -5,6 +6,7 @@ import { Constructor, ImplementationType } from '../types';
 function singleton<TClass extends Constructor>(constructor: TClass): TClass {
   (constructor as any)[singleton_s] = true;
   InheritancePreserver.constructorModified(constructor);
+  di(constructor);
 
   return constructor;
 }